requirements.txt
opencv-python
numpy
mss
pyautogui
Pillow
keyboard
scikit-image
pyserial
